+++ src/aac_datasets/__init__.py
@@ -1,9 +1,17 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
 
-"""
-Audio Captioning Datasets package.
-"""
+"""Audio Captioning datasets for PyTorch.  """
+
+
+__author__ = "Etienne Labbé (Labbeti)"
+__author_email__ = "labbeti.pub@gmail.com"
+__license__ = "MIT"
+__maintainer__ = "Etienne Labbé (Labbeti)"
+__name__ = "aac-datasets"
+__status__ = "Development"
+__version__ = "0.4.1"
+
 
 from .datasets.audiocaps import AudioCaps
 from .datasets.clotho import Clotho
@@ -19,10 +27,15 @@ from .utils.paths import (
 )
 
 
-__author__ = "Etienne Labbé (Labbeti)"
-__author_email__ = "labbeti.pub@gmail.com"
-__license__ = "MIT"
-__maintainer__ = "Etienne Labbé (Labbeti)"
-__name__ = "aac-datasets"
-__status__ = "Development"
-__version__ = "0.4.0"
+__all__ = [
+    "AudioCaps",
+    "Clotho",
+    "MACS",
+    "WavCaps",
+    "get_default_ffmpeg_path",
+    "get_default_root",
+    "get_default_ytdl_path",
+    "set_default_ffmpeg_path",
+    "set_default_root",
+    "set_default_ytdl_path",
+]