+++ src/aac_datasets/datasets/base.py
@@ -21,8 +21,13 @@ from typing import (
 import torchaudio
 import tqdm
 
-from torchaudio.backend.common import AudioMetaData
 from typing_extensions import TypedDict
+
+try:
+    # To support torchaudio >=2.1.0
+    from torchaudio import AudioMetaData  # type: ignore
+except ImportError:
+    from torchaudio.backend.common import AudioMetaData
 
 from torch import Tensor
 from torch.utils.data.dataset import Dataset
@@ -101,12 +106,22 @@ class AACDataset(Generic[ItemType], Dataset[ItemType]):
         return self._flat_captions
 
     @property
+    def num_columns(self) -> int:
+        """Number of columns in the dataset."""
+        return len(self.column_names)
+
+    @property
+    def num_rows(self) -> int:
+        """Number of rows in the dataset (same as len())."""
+        return len(self)
+
+    @property
     def raw_data(self) -> Dict[str, List[Any]]:
         return self._raw_data
 
     @property
     def shape(self) -> Tuple[int, int]:
-        """The shape of the dataset."""
+        """Shape of the dataset (number of columns, number of rows)."""
         return len(self), len(self.column_names)
 
     @property