+++ src/aac_datasets/datasets/macs.py
@@ -26,7 +26,7 @@ from torch.hub import download_url_to_file
 from typing_extensions import TypedDict
 
 from aac_datasets.datasets.base import AACDataset, DatasetCard
-from aac_datasets.utils.download import validate_file
+from aac_datasets.utils.download import hash_file
 from aac_datasets.utils.paths import _get_root
 
 
@@ -147,14 +147,19 @@ class MACS(AACDataset[MACSItem]):
 
         if download:
             _prepare_macs_dataset(
-                root,
-                subset,
-                verbose,
-                MACS.FORCE_PREPARE_DATA,
-                MACS.VERIFY_FILES,
-                MACS.CLEAN_ARCHIVES,
+                root=root,
+                subset=subset,
+                verbose=verbose,
+                force=MACS.FORCE_PREPARE_DATA,
+                verify_files=MACS.VERIFY_FILES,
+                clean_archives=MACS.CLEAN_ARCHIVES,
             )
-        raw_data, annotator_id_to_competence = _load_macs_dataset(root, subset, verbose)
+
+        raw_data, annotator_id_to_competence = _load_macs_dataset(
+            root=root,
+            subset=subset,
+            verbose=verbose,
+        )
 
         audio_dpath = _get_audio_dpath(root)
         size = len(next(iter(raw_data.values())))
@@ -395,9 +400,12 @@ def _prepare_macs_dataset(
 
         if verify_files:
             hash_value = file_info["hash_value"]
-            valid = validate_file(fpath, hash_value, hash_type="md5")
-            if not valid:
-                raise RuntimeError(f"Invalid checksum for file {fname}.")
+            file_hash_value = hash_file(fpath, hash_type="md5")
+            if file_hash_value != hash_value:
+                raise RuntimeError(
+                    f"Invalid checksum for file '{fname}'. (expected md5 checksum '{hash_value}' but found '{file_hash_value}')\n"
+                    f"Please try to remove manually the file '{fpath}' and rerun MACS download."
+                )
             elif verbose >= 2:
                 pylog.debug(f"File '{fname}' has a valid checksum.")
 
@@ -426,9 +434,12 @@ def _prepare_macs_dataset(
 
         if verify_files:
             hash_value = file_info["hash_value"]
-            valid = validate_file(zip_fpath, hash_value, hash_type="md5")
-            if not valid:
-                raise RuntimeError(f"Invalid checksum for file {zip_fname}.")
+            file_hash_value = hash_file(zip_fpath, hash_type="md5")
+            if file_hash_value != hash_value:
+                raise RuntimeError(
+                    f"Invalid checksum for file '{zip_fname}'. (expected md5 checksum '{hash_value}' but found '{file_hash_value}')\n"
+                    f"Please try to remove manually the file '{zip_fpath}' and rerun MACS download."
+                )
             elif verbose >= 2:
                 pylog.debug(f"File '{zip_fname}' has a valid checksum.")